Clean up for kernel 5.10.x
- Add individual kernel-modules directories for each native kernel, i.e. v7 and v7l
- Change kernel version to 5.10.20
- Supress error messages when optional files cannot be copied
- Use "Image" instead of "zImage" for 64bit kernels

+++ build.sh
@@ -17,7 +17,7 @@
 DEBUG="0"
 
 ## Version strings:
-VERSION="5.10.40"
+VERSION="5.10.20"
 BUILD="1"
 unset V6_VERSION V7_VERSION V7L_VERSION V8_VERSION V8L_VERSION
 ## Comment out those you don't want to build
@@ -173,7 +173,7 @@ EOF
 function clean() {
     if [ ! -d $KERNEL_SRC_DIR ]; then
         setup_repos
-    fi	
+    fi
     clean_kernel_src_dir
     echo "**** Cleaning up kernel working dirs ****"
     if [ -d $KERNEL_OUT_DIR_V6 ]; then
@@ -231,12 +231,17 @@ function clean() {
             echo $version > $KERNEL_OUT_DIR_V8L\/\.version
         fi
     fi
-    echo "**** Kernel source directories cleaned up ****"
     if [ -d $KERNEL_HEADERS_OUT_DIR ]; then
         rm -rf $KERNEL_HEADERS_OUT_DIR
         mkdir -p $KERNEL_HEADERS_OUT_DIR
         chown $SUDO_USER:$SUDO_USER $KERNEL_HEADERS_OUT_DIR
     fi
+    if [ -d $KERNEL_MOD_DIR ]; then
+        rm -rf $KERNEL_MOD_DIR
+        mkdir -p $KERNEL_MOD_DIR
+        chown $SUDO_USER:$SUDO_USER $KERNEL_MOD_DIR
+    fi
+    echo "**** Kernel working directories cleaned up ****"
     cd -
 }
 
@@ -254,7 +259,7 @@ function clone_source() {
     echo "**** CLONING to ${REPO_ROOT}${GIT_REPO} ****"
     echo "REPO: ${GIT_REPO}"
     echo "BRANCH: ${GIT_BRANCH}"
-    git clone --recursive https://github.com/${GIT_REPO} $KERNEL_SRC_DIR 
+    git clone --recursive https://github.com/${GIT_REPO} $KERNEL_SRC_DIR
 }
 
 function setup_repos(){
@@ -282,7 +287,7 @@ function setup_repos(){
     fi
     if [ ! -d $HEAD_SRC_DIR ]; then
 	mkdir -p $HEAD_SRC_DIR
-    fi	
+    fi
 
 #    if [ ! -d $TOOLS_DIR ]; then
 #        echo "**** CLONING TOOL REPO ****"
@@ -380,7 +385,7 @@ function update_kernel_source(){
     git pull
     git submodule update --init
 
-    ##get_4d_obj 
+    ##get_4d_obj
     cd -
 }
 
@@ -388,27 +393,33 @@ function prep_kernel_out_dir() {
     kernel_out_dir=$1
     printf "\n**** PREPARING WORKING DIRECTORY $kernel_out_dir ****\n"
     if [ ! -d $kernel_out_dir ]; then
-        mkdir -p $kernel_out_dir 
+        mkdir -p $kernel_out_dir
     fi
     ## The following workarounds are required for a successful compilation
     ## Copy 4D pre-compiled object files
     ##get_4d_obj
     if [ ! -d $kernel_out_dir/drivers/video/4d-hats ]; then
-        mkdir -p $kernel_out_dir/drivers/video/4d-hats 
+        mkdir -p $kernel_out_dir/drivers/video/4d-hats
     fi
-    ##cp $KERNEL_SRC_DIR/drivers/video/4d-hats/compress-*.o $kernel_out_dir/drivers/video/4d-hats/
+    if [ -f $KERNEL_SRC_DIR/drivers/video/4d-hats/compress-*.o ]; then
+        cp $KERNEL_SRC_DIR/drivers/video/4d-hats/compress-*.o $kernel_out_dir/drivers/video/4d-hats/
+    fi
     ## Copy rtl8812au files required for the compilation
     ## The path in the driver source has to be fixed before we can renove this workaround
     if [ ! -d $kernel_out_dir/drivers/net/wireless/realtek/rtl8812au/hal/phydm ]; then
         mkdir -p $kernel_out_dir/drivers/net/wireless/realtek/rtl8812au/hal/phydm
     fi
-    cp -rf $KERNEL_SRC_DIR/drivers/net/wireless/realtek/rtl8812au/hal/phydm/phydm.mk $kernel_out_dir/drivers/net/wireless/realtek/rtl8812au/hal/phydm/
+    if [ -f $KERNEL_SRC_DIR/drivers/net/wireless/realtek/rtl8812au/hal/phydm/phydm.mk ]; then
+        cp -rf $KERNEL_SRC_DIR/drivers/net/wireless/realtek/rtl8812au/hal/phydm/phydm.mk $kernel_out_dir/drivers/net/wireless/realtek/rtl8812au/hal/phydm/
+    fi
     ## Copy rtl8192eu files required for the compilation
     ## The path in the driver source has to be fixed before we can renove this workaround
     if [ ! -d $kernel_out_dir/drivers/net/wireless/realtek/rtl8192eu/hal/phydm ]; then
         mkdir -p $kernel_out_dir/drivers/net/wireless/realtek/rtl8192eu/hal/phydm
     fi
-    cp -rf $KERNEL_SRC_DIR/drivers/net/wireless/realtek/rtl8192eu/hal/phydm/phydm.mk $kernel_out_dir/drivers/net/wireless/realtek/rtl8192eu/hal/phydm/
+    if [ -f $KERNEL_SRC_DIR/drivers/net/wireless/realtek/rtl8192eu/hal/phydm/phydm.mk ]; then
+        cp -rf $KERNEL_SRC_DIR/drivers/net/wireless/realtek/rtl8192eu/hal/phydm/phydm.mk $kernel_out_dir/drivers/net/wireless/realtek/rtl8192eu/hal/phydm/
+    fi
 }
 
 function make_v6() {
@@ -525,7 +536,8 @@ function make_v8() {
     echo "**** SAVING A COPY OF YOUR v8 CONFIG TO $KERNEL_BUILDER_DIR/configs/re4son_pi8_defconfig ****"
     cp -f $KERNEL_OUT_DIR_V8/.config $KERNEL_BUILDER_DIR/configs/re4son_pi8_defconfig
     echo "**** COMPILING v8 KERNEL ****"
-    make ARCH=arm64 CROSS_COMPILE=${CCPREFIX} O=$KERNEL_OUT_DIR_V8 -C $KERNEL_SRC_DIR -j${NUM_CPUS} 
+    ## zImage is not supported for 64bit architecture so we build Image
+    make ARCH=arm64 CROSS_COMPILE=${CCPREFIX} O=$KERNEL_OUT_DIR_V8 -C $KERNEL_SRC_DIR -j${NUM_CPUS} Image modules dtbs
     make ARCH=arm64 CROSS_COMPILE=${CCPREFIX} O=$KERNEL_OUT_DIR_V8 -C $KERNEL_SRC_DIR INSTALL_MOD_PATH=${MOD_DIR} -j${NUM_CPUS} modules_install
     ## Name the kernel "kernel8-alt.img" for now to prevent it from automatically being loaded
     ## To use it, just rename it to kernel8.img on the device
@@ -559,7 +571,8 @@ function make_v8l() {
     echo "**** SAVING A COPY OF YOUR v8l CONFIG TO $KERNEL_BUILDER_DIR/configs/re4son_pi8l_defconfig ****"
     cp -f $KERNEL_OUT_DIR_V8L/.config $KERNEL_BUILDER_DIR/configs/re4son_pi8l_defconfig
     echo "**** COMPILING v8l KERNEL ****"
-    make ARCH=arm64 CROSS_COMPILE=${CCPREFIX} O=$KERNEL_OUT_DIR_V8L -C $KERNEL_SRC_DIR -j${NUM_CPUS} 
+    ## zImage is not supported for 64bit architecture so we build Image
+    make ARCH=arm64 CROSS_COMPILE=${CCPREFIX} O=$KERNEL_OUT_DIR_V8L -C $KERNEL_SRC_DIR -j${NUM_CPUS} Image modules dtbs
     make ARCH=arm64 CROSS_COMPILE=${CCPREFIX} O=$KERNEL_OUT_DIR_V8L -C $KERNEL_SRC_DIR INSTALL_MOD_PATH=${MOD_DIR} -j${NUM_CPUS} modules_install
     ## Name the kernel "kernel8-alt.img" for now to prevent it from automatically being loaded
     ## To use it, just rename it to kernel8.img on the device
@@ -601,8 +614,11 @@ function make_native_v6() {
     rm -f ${MOD_DIR}/lib/modules/*/source
     ## Copy our modules across
     cp -r ${MOD_DIR}/lib/* ${PKG_DIR}
-    ## Copy away the module dir so we cak use it for compiling drivers if we want
-    cp -r ${MOD_DIR}/lib/modules/*/* ${KERNEL_MOD_DIR}/
+    ## Copy away the module dir so we can use it for compiling drivers if we want
+    if [ ! -d ${KERNEL_MOD_DIR}/v6 ]; then
+	mkdir ${KERNEL_MOD_DIR}/v6
+    fi
+    cp -r ${MOD_DIR}/lib/modules/*Re4son+/* ${KERNEL_MOD_DIR}/v6/
     ## Copy our Module.symvers across
     mkdir -p $PKG_DIR/headers/usr/src/linux-headers-$UNAME_STRING/
     cp $KERNEL_OUT_DIR_V6/Module.symvers $PKG_DIR/headers/usr/src/linux-headers-$UNAME_STRING/
@@ -638,8 +654,11 @@ function make_native_v7() {
     rm -f ${MOD_DIR}/lib/modules/*-v7+/source
     ## Copy our modules across
     cp -r ${MOD_DIR}/lib/* ${PKG_DIR}
-    ## Copy away the module dir so we cak use it for compiling drivers if we want
-    cp -r ${MOD_DIR}/lib/modules/*/* ${KERNEL_MOD_DIR}/
+    ## Copy away the module dir so we can use it for compiling drivers if we want
+    if [ ! -d ${KERNEL_MOD_DIR}/v7 ]; then
+	mkdir ${KERNEL_MOD_DIR}/v7
+    fi
+    cp -r ${MOD_DIR}/lib/modules/*v7+/* ${KERNEL_MOD_DIR}/v7/
     ## Copy our Module.symvers across
     mkdir -p $PKG_DIR/headers/usr/src/linux-headers-$UNAME_STRING7/
     cp $KERNEL_OUT_DIR_V7/Module.symvers $PKG_DIR/headers/usr/src/linux-headers-$UNAME_STRING7/
@@ -672,8 +691,11 @@ function make_native_v7l() {
     rm -f ${MOD_DIR}/lib/modules/*-v7l+/source
     ## Copy our modules across
     cp -r ${MOD_DIR}/lib/* ${PKG_DIR}
-    ## Copy away the module dir so we cak use it for compiling drivers if we want
-    cp -r ${MOD_DIR}/lib/modules/*/* ${KERNEL_MOD_DIR}/
+    ## Copy away the module dir so we can use it for compiling drivers if we want
+    if [ ! -d ${KERNEL_MOD_DIR}/v7l ]; then
+	mkdir ${KERNEL_MOD_DIR}/v7l
+    fi
+    cp -r ${MOD_DIR}/lib/modules/*v7l+/* ${KERNEL_MOD_DIR}/v7l/
     ## Copy our Module.symvers across
     mkdir -p $PKG_DIR/headers/usr/src/linux-headers-$UNAME_STRING7L/
     cp $KERNEL_OUT_DIR_V7L/Module.symvers $PKG_DIR/headers/usr/src/linux-headers-$UNAME_STRING7L/
@@ -696,20 +718,24 @@ function make_native_v8() {
     echo "**** SAVING A COPY OF YOUR v8 CONFIG TO $KERNEL_BUILDER_DIR/configs/re4son_pi8_defconfig ****"
     cp -f $KERNEL_OUT_DIR_V8/.config $KERNEL_BUILDER_DIR/configs/re4son_pi8_defconfig
     echo "**** COMPILING v8 KERNEL ****"
-    make O=$KERNEL_OUT_DIR_V8 -C $KERNEL_SRC_DIR -j${NUM_CPUS}
+    ## zImage is not supported for 64bit architecture so we build Image
+    make O=$KERNEL_OUT_DIR_V8 -C $KERNEL_SRC_DIR -j${NUM_CPUS} Image modules dtbs
     make O=$KERNEL_OUT_DIR_V8 -C $KERNEL_SRC_DIR INSTALL_MOD_PATH=${MOD_DIR} -j${NUM_CPUS} modules_install
     ## Name the kernel "kernel8-alt.img" for now to prevent it from automatically being loaded
     ## To use it, just rename it to kernel8.img on the device
     ## Apparently we don't use mkknlimg anymore:
     ## https://github.com/raspberrypi/linux/issues/3249#issuecomment-534134438
-    cp -f $KERNEL_OUT_DIR_V8/arch/arm64/boot/zImage $PKG_DIR/boot/kernel8-alt.img
+    cp -f $KERNEL_OUT_DIR_V8/arch/arm64/boot/Image $PKG_DIR/boot/kernel8-alt.img
     ## Remove symbolic links to non-existent headers and sources
     rm -f ${MOD_DIR}/lib/modules/*-v8+/build
     rm -f ${MOD_DIR}/lib/modules/*-v8+/source
     ## Copy our modules across
     cp -r ${MOD_DIR}/lib/* ${PKG_DIR}
-    ## Copy away the module dir so we cak use it for compiling drivers if we want
-    cp -r ${MOD_DIR}/lib/modules/*/* ${KERNEL_MOD_DIR}/
+    ## Copy away the module dir so we can use it for compiling drivers if we want
+    if [ ! -d ${KERNEL_MOD_DIR}/v8 ]; then
+	mkdir ${KERNEL_MOD_DIR}/v8
+    fi
+    cp -r ${MOD_DIR}/lib/modules/*v8+/* ${KERNEL_MOD_DIR}/v8/
     ## Copy our Module.symvers across
     mkdir -p $PKG_DIR/headers/usr/src/linux-headers-$UNAME_STRING8/
     cp $KERNEL_OUT_DIR_V8/Module.symvers $PKG_DIR/headers/usr/src/linux-headers-$UNAME_STRING8/
@@ -732,20 +758,24 @@ function make_native_v8l() {
     echo "**** SAVING A COPY OF YOUR v8l CONFIG TO $KERNEL_BUILDER_DIR/configs/re4son_pi8l_defconfig ****"
     cp -f $KERNEL_OUT_DIR_V8L/.config $KERNEL_BUILDER_DIR/configs/re4son_pi8l_defconfig
     echo "**** COMPILING v8l KERNEL ****"
-    make O=$KERNEL_OUT_DIR_V8L -C $KERNEL_SRC_DIR -j${NUM_CPUS}
+    ## zImage is not supported for 64bit architecture so we build Image
+    make O=$KERNEL_OUT_DIR_V8L -C $KERNEL_SRC_DIR -j${NUM_CPUS} Image modules dtbs
     make O=$KERNEL_OUT_DIR_V8L -C $KERNEL_SRC_DIR INSTALL_MOD_PATH=${MOD_DIR} -j${NUM_CPUS} modules_install
     ## Name the kernel "kernel8l-alt.img" for now to prevent it from automatically being loaded
     ## To use it, just rename it to kernel8l.img on the device
     ## Apparently we don't use mkknlimg anymore:
     ## https://github.com/raspberrypi/linux/issues/3249#issuecomment-534134438
-    cp -f $KERNEL_OUT_DIR_V8L/arch/arm64/boot/zImage $PKG_DIR/boot/kernel8l-alt.img
+    cp -f $KERNEL_OUT_DIR_V8L/arch/arm64/boot/Image $PKG_DIR/boot/kernel8l-alt.img
     ## Remove symbolic links to non-existent headers and sources
     rm -f ${MOD_DIR}/lib/modules/*-v8l+/build
     rm -f ${MOD_DIR}/lib/modules/*-v8l+/source
     ## Copy our modules across
     cp -r ${MOD_DIR}/lib/* ${PKG_DIR}
-    ## Copy away the module dir so we cak use it for compiling drivers if we want
-    cp -r ${MOD_DIR}/lib/modules/*/* ${KERNEL_MOD_DIR}/
+    ## Copy away the module dir so we can use it for compiling drivers if we want
+    if [ ! -d ${KERNEL_MOD_DIR}/v8l ]; then
+	mkdir ${KERNEL_MOD_DIR}/v8l
+    fi
+    cp -r ${MOD_DIR}/lib/modules/*v8l+/* ${KERNEL_MOD_DIR}/v8l/
     ## Copy our Module.symvers across
     mkdir -p $PKG_DIR/headers/usr/src/linux-headers-$UNAME_STRING8L/
     cp $KERNEL_OUT_DIR_V8L/Module.symvers $PKG_DIR/headers/usr/src/linux-headers-$UNAME_STRING8L/
